requirements.txt
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-Env==2.0.0
future==0.18.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
six==1.16.0
torch==1.5.1+cpu
torchvision==0.6.1+cpu
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
waitress==2.0.0

-f https://download.pytorch.org/whl/torch_stable.html